+++ src/components/AppContainer.js
@@ -1,0 +1,16 @@
+import HomeScreen from './HomeScreen';
+import DetailsScreen from './DetailsScreen';
+
+import { createStackNavigator, createAppContainer } from "react-navigation";
+
+const AppNavigator = createStackNavigator(
+  {
+    Home: HomeScreen,
+    Details: DetailsScreen
+  },
+  {
+    initialRouteName: "Home"
+  }
+);
+
+const AppContainer = createAppContainer(AppNavigator);
